Moved fixtures to data
