+++ data/fixtures/output-markdown/resource-table.md
