+++ data/fixtures/output-markdown/resource.md
