+++ data/fixtures/output-unzipped/xl/drawings/vmlDrawing1.vml
