+++ data/fixtures/output-unzipped/xl/sharedStrings.xml
