+++ data/fixtures/output-unzipped/xl/theme/theme1.xml
